input-surface: Move debug info to separate widget
Move the surface that keeps the debug information (which was
phosh-osk-stub's initial purpose) into a separate widget. This makes
pos-input-surface focus more in input and also allows us to add new
things to the debug surface more easily without getting into conflict
with e.g. the completion code.

We keep the input-surface's callbacks needed for completion in place.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>