input-surface: Allow to toggle completion via menu
We do so by using a GPropertyAction that binds the `win.word-completion`
action to the `completion-enabled` property.

Since GPropertyAction holds a ref on the object (PosInputSurface) we
need to connect to the `destroy` signal in order to break the reference
cycle when calling `gtk_wiget_destroy` on PosInputSurface. Otherwise the
input surface would never get disposed since the action_map holds a
reference on the GPropertyAction which in turn holds a reference on the
input surface.

Since we don't want to have `destroy` and `dispose` we move the cleanup
of the hash map there too.

With GTK4 we can move to `gtk_widget_class_install_property_action` to
get rid of this.

Signed-off-by: Guido Günther <agx@sigxcpu.org>