input-surface: Restore keymap with slight delay
When closing the popup apply a slight delay before restoring the keymap
as otherwise the symbol sent to the popup might get applied to the new
keymap set (rather than the one set for the popup).

The virtual-keyboard requests are sent from pos int the right ordering
and the compositor gets them correctly too so it fails within the X11
bits.

Closes: https://gitlab.gnome.org/guidog/phosh-osk-stub/-/issues/22