osk-stub: Add minimal GObject
This will allow us to remove the static variables in there and hence
make the surface creation easier to track.

As a start we handle the main loop from there which fixes the issue
that we passed a NULL `loop` to `pos_session_register`.

We don't use `G(tk)Application` as with that class we can't reown the
DBus name with `--allow-replacement` as the primary instance is being
stopped. We might revisit that after cleaning up more.

Signed-off-by: Guido Günther <agx@sigxcpu.org>